// frontend/src/enums/paths.ts
export enum Paths {
    Dashboard = "dashboard",
    PersonalCalendar = "personal-calendar",
    CompanyBackground = "company-background",
    Notifications = "notifications",
    Events = "events",
    Assets = "assets",
    OrganisationalStructure = "organisational-structure",
    Salaries = "salaries",
    DayOff = "day-off",
    Promotion = "promotion",
    Recruitment = "recruitment",
    Employment = "employment",
    Dismissed = "dismissed",
    Login = "/"
}